Bug 2759: moving more code from browser_xbl_temp.js to sb-tabbrowser.xml
This patch moves a bunch of code from browser_xbl_temp.js to sb-tabbrowser.xml
and does a few cleanups in the process

* the use of the browser-status-filter is partially reintroduced, and the error
on shutdown that was introduced by its removal has been removed
* the servicepane now knows about the browser through an XML attribute, not by
assigning to a field.
* from that the way that the first page loads at startup changed (it's done in
the servicepane constructor not as a side effect of setting
servicepane.browser)
* the hack to enable scrollbars moved from the servicepane into the
sb-tabbrowser - if it's going to be necessary, that's where it should be.


git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@3845 5aad4ff5-4b58-4159-ba3f-0e6328df4cad