For bug 2845. The general effect of this patch is to add tabbed browsing.
Part of this involves convincing the mozilla tabbrowser to play nicely. To this end I added a bunch of default prefs, created a global getBrowser() method (like firefox has) and commented out a tiny bit of tabbrowser.xml that referenced menu items that we don't have by id.

In the sb-tabbrowser, tab switching is now tracked so we can re-trigger the media scan. There's a new loadURIFromEvent that will open a new tab (or not) based on keyboard modifiers and mouse buttons. The servicepane uses this interface now.

The location bar now has autocomplete enabled, and is a web progress listener for the tabbrowser so it can update when a new page loads or when the user switches tabs. This moves a bunch of code out of the onLocationChange on sb-tabbrowser.

The browser link context menu got a new "Open in New Tab" option that does what you'd expect.

There was some code left over from the pre-bubble web playlist callout. It's gone now.


git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@3951 5aad4ff5-4b58-4159-ba3f-0e6328df4cad