bug 2962 add a mechanism to localize service pane node items
So while working on drag and drop I switched the tree type and broke the
localization code I'd written. That code sucked anyway so it's for the best.

This introduces a proxy rdf datasource that translates specific node properties
(currently just NC:Name). I also fixed a typo with the dnd property names and
introduced some test cases.



git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@4412 5aad4ff5-4b58-4159-ba3f-0e6328df4cad