Bug 3782
This change prevents the onLocationChange and onStateChange handlers in tabs
running if the progress is not for the top-level window. This fixes some IFRAME
problems. Also it checks to see if the right document is loading based on
linkedBrowser.contentDocument.documentURI not aRequest.name as the latter is
incorrect when loading chrome (at least in debug builds on Linux).

This allows library views to be played with the play button.


git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@5347 5aad4ff5-4b58-4159-ba3f-0e6328df4cad