Bug 4400 - First commit. Fixed 'window_parent' and 'frame_mini' references.
I eat that error because we're in a callback function and
we need to return that -1.  Also, miniPlayer.js isn't as
necessary to cleanup, right?  Minmax stuff should be specific
to the layout being written.

Hopefully, someday, we won't have to do our own minmax handling.
We fixed it on the branch, but I don't think it ever got migrated.

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@6190 5aad4ff5-4b58-4159-ba3f-0e6328df4cad