bug 5689 bug 3278 bug 5690 - play and skip count and time This adds tracking of the above 4 pieces of metadata.
To test:
begin playback from library
select the columns from the picker for these fields
  (Number of Times Played, Number of Times Skipped,
   Last Time Skipped, Last Time Played)

Play should be counted when:
a track reaches the halfway point without dragging the position
if the position is dragged to the end and the track is let to finish
if 4 minutes of time in playback OR half the duration of 
the track is spend in playback. You could let 2 minutes play on a 5 minute
track and then scan back to 1 minute twice, letting it run to 2 minutes each time
To test the 4 minute piece you need a track > 8min long.
The Last Played time should be updated at this point as well.
The Last Played time should also update when a track reaches the
end and advances to the next track.

Skip should be counted when:
Next or Prev is clicked in the UI when the following conditions exist:
1) More than 2 seconds of playback have occured, regardless of position in track
AND
2) Less than 15 seconds of playback has occured, regardless of position in track

We don't count a track as skipped if the user 


git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@7272 5aad4ff5-4b58-4159-ba3f-0e6328df4cad