Bug 4719: Rework the "Startup" prefs section into a "Home Page" one.
So, we don't actually want to allow users to decide if they want their tabs restored or not. We want to always restore all their tabs, so we need to hide a bunch of UI. Also, we don't want to let them pick their home page from bookmarks (which never worked anyway). There are bigger, better, long-term plans for this.

There will be another patch later to make "Use Current Page" actually work.


git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@7326 5aad4ff5-4b58-4159-ba3f-0e6328df4cad