Bug 5981: Fix navigation in sub-frames
So the problem was that we were treating all links with no target as links to be loaded at the top-level. This was wrong in the frames case. In the frames case if no frame was specified you want to load in the current frame. I found (after re-reading Firefox's browser.js) that we can just return true from our onclick handler to have the default behavior take place.

So:
* I modified sbtabbrowser._checkTarget to return null if no target was passed in.
* I made loadURI explicitly treat a blank target as _top
* I made the onclick handler not call loadURI, but return true when there's no target

It seems to work. I have a test page at: http://scratch.ianloic.com/songbird/frames/
It has four frames named
(a) (b)
(c) (d)

and four pages. Each page has links to each other page in each frame and without a target. This way you can click around and show that targetting (or lack therof) works correctly.


git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@8091 5aad4ff5-4b58-4159-ba3f-0e6328df4cad