Bug 8321 - [mtp] Sync device on playlist change
Minor fixup of the previous patch for Bug 8321. This fixes the issue where I was turning off listeners for playlists right after I set them up. The stop call needed to occur before I setup the listeners so we were only listening to the playlists we're interested in.

This also has a minor fix to quite a warning where sbLocalDatabaseMediaListBase::GetItemsByProperties was dumping to console when there were no items.

Also changed sbLocalDatabaseMediaItem::GetContentLength not to return an error if the content property is blank, it will return 0

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@9431 5aad4ff5-4b58-4159-ba3f-0e6328df4cad