Bug 10178 - [csperf]Implement changes to the library to get the UI up faster for larger libraries
This caches the partial array enumeration that occurs in the media lists. Whenever a single value property enumeration occurs it see's if the key and value are the same and if so uses the previously cached GUID Array. If it's different then blows away the GUID array and caches the newly created one.

r=erikstats


git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@9921 5aad4ff5-4b58-4159-ba3f-0e6328df4cad