Bug 10249 - [mtp] manually managed library disallows metadata editing (is read only)
Turned out the GetUserEditable implementation was at fault. Apparently logic was added to make non-local items read-only. So I added an exception for x-mtp schemes.

Also a small fix piggy backed onto this that hides hidden playlists from the list of playlists lists to sync on the device summary page.

r=mike-s

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@10025 5aad4ff5-4b58-4159-ba3f-0e6328df4cad