Bug 10820 - problems with list of playlists in device page
The mCachedPartialArray wasn't being invalidated when the mFullArray was. This caused it to get out of sync. Listening for the invalidate and then invaliding the cached array keeps everything in sync.

This also moved mFullArray to private so that it can't be replaced without the base class's knowledge.

As well as the bug this also quiets the GetMediaItem passed a bad guid assert.

r=erikstaats

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@10305 5aad4ff5-4b58-4159-ba3f-0e6328df4cad