Bug 11440 - CSPerfEndEULA isn't getting recorded
This fixes the first run wizard so we don't start the restart before we have finalized the wizard. That was the real issue behind the missing time stamp.

I also changed the CSPerfEndEULA start and stop in the same place, as if we restart at the end of the wizard we don't get that timestamp.

This also fixes an issue in the FolderUtils.jsm where it was looking up Home and getting an exception. Changed to check for existence first and leave null if Home isn't found.

r=pvh

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@10423 5aad4ff5-4b58-4159-ba3f-0e6328df4cad