Bug 11440 - CSPerfEndEULA isn't getting recorded
This is another part of the fix for this bug. In the end we really need to have the timing service append output to the log and not truncate it. This is due to the restarts that occur during startup.

Also the file wasn't being closed afterwards which caused problems on subsequent opens.

r=mook

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@10474 5aad4ff5-4b58-4159-ba3f-0e6328df4cad