Bug 12947 - Playlist widget doesn't redraw when it should
Moves the media list view loading logic during first run from the job listener to the main thread. This fixes issues that occur with the media list view after. Specifically deleting all doesn't update the window and leaves all the items there, which in turns cause JS exceptions. Ideally this should use a proxy object, but I didn't see an existing example and this is was easy and safe to get in.

r=erikstaats

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@11452 5aad4ff5-4b58-4159-ba3f-0e6328df4cad