Bug 13222 - [topcrash] sbFixedInterfaceCache<nsStringHashKey, sbLocalDatabaseResourcePropertyBag>::Put(nsAString const&, sbLocalDatabaseResourcePropertyBag*)
Prevent putting nulls in the property bag cache. Also changed the sbFixedInterfaceCache so that it handles nulls ok, since it's a generic utility container it should crap out on nulls.

Also updated the library guid logic in RetrieveProperties. The new prepared statement logic allows this to be a lot cleaner than it was before.

r=mattc

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@11550 5aad4ff5-4b58-4159-ba3f-0e6328df4cad