Bug 13287 - [♪photo?] media items in the library go transparent all of the sudden
This adds logic to handle an error path. The original logic didn't handle a ordinal as a sort spect when the base table was something other than simple_media_lists. This switches the sort to created in that case. There's another bug at work here, but this handles this odd state more gracefully.

r=pvh

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@11750 5aad4ff5-4b58-4159-ba3f-0e6328df4cad