Bug 15342 - [watchfolders] Crashes occurs when accessing mTimer on Cleanup
Move the Cleanup call to be on the thread so that freeing of resources doesn't encounter threading issues.

Fix the file system events unit test so it handles failure on removing the directory and retries a few times. Before this was failing and the test would hang as testFinished would never get called.

r=kreeger

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@12847 5aad4ff5-4b58-4159-ba3f-0e6328df4cad