Bug 16020 - [shakur] [perf] Adapt the iTunes importer to the new C++ API
C++ implementation of iTunes import.

This also changes the sbiTunesXMLParser to use a stream pump so that it can run on the main thread and not block the UI.

not currently hooked up to build. JS code still be usinged. There's issues with the UI updating. Doesn't seem to play well with the modal dialog. Everything is operational except the UI progress issue. We'll be dealing with that in a bug.

r=mikes

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@13468 5aad4ff5-4b58-4159-ba3f-0e6328df4cad