Bug 16124 - [shakur][agent] Agent should quit when the shutdown file is present
This finishes off the Windows implementation for watching for the shutdown file. The biggest problem was the way the error was getting built. The compiler chose the boolean constructor which just made it look like a non-error. Also renamed the STRINGIT to STRINGIZE and moved them to base class so we just have one definition.

r=nick

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@13518 5aad4ff5-4b58-4159-ba3f-0e6328df4cad