Bug 16575: iTunes playlist is being imported twice
I made the iTunes GUID a regular property because I needed it elsewhere.

Added logic so we don't import back in the Songbird playlists that were
previous exported.

No longer import iTunes folders as playlists.

Changed the media list creation so that we pass in properties instead of
creating then setting the properties. This allows the export to know what it's
looking at instead of just seeing a blank no named media list.

Fixed a problem in the iTunes agent's main function. The local error object
would assert when exiting due to another process running. This exit path would
never do anything with it, so it would fire the assert.

r=nick

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@13777 5aad4ff5-4b58-4159-ba3f-0e6328df4cad