Bug 19455 - [videoplayback] Video viewing crashes occasionally
Use var instead of let, as let has issues with nested callback functions.

The use of let was causing JS asserts. JS was asked to add a slot to a dead or otherwise invalid JS object. The assignment of the let variable apparently referenced the old JS function scope which had been collected.

r=mook

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@16986 5aad4ff5-4b58-4159-ba3f-0e6328df4cad