Bug 19850 - [sync] video files are syncing if autosync all music is selected
Changed sync logic so that it uses a smart media list and some post filter to handle syncing different media types.

Changed the sync management methods in sbIDeviceLibrary to simplify and reflect what we're really doing with syncing.

Fixed up the UI code to match this and address some spots that needed tweaking to account for sync'ing different media types.

r=stevo, aus

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@17386 5aad4ff5-4b58-4159-ba3f-0e6328df4cad