Bug 22961 - [topcrash] sbBaseDevice::ShutdownRequestThread()
This is just the new code for the request thread queue logic. Once the rest of the reviews are in place this will get hooked up. This code was extracted and refactored from the sbBaseDevice class. The goal is to separate and simplify the request processing logic so that the device disconnect logic as well as other aspects are easier to manage. The old logic it was very difficult to follow and understand all the issues between thread shutdown and device disconnect. And many devices had their own nuances to this as well.

r=msmith

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@20613 5aad4ff5-4b58-4159-ba3f-0e6328df4cad