Bug 22961 - [topcrash] sbBaseDevice::ShutdownRequestThread()
This is the changes for the unit test and a couple of changes to the request queue and device that were issues in the way the mock device was used. These issues shouldn't have an impact on the real devices, they mostly centered around the fact that in the unit test the only thing that holds onto the device is the request thread when shutting down.

r=msmith

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@20842 5aad4ff5-4b58-4159-ba3f-0e6328df4cad