Bug 22666 - [sync2] Items dragged from desktop onto device library re-"added"
1. The external import code did not filter dupes out based on origin url. It
only looked at content URL which is not enough when importing to a non-main
library. All the changes for this are in sbDirectoryImportService.js.

2. There was a bug in
sbSimpleMediaListInsertingEnumerationListener::OnEnumerationEnd that didn't
properly handle all three scenarios. 1. Item exists in main library, 2. Item
exists in target. 3. Item doesn't exist. It was treating 2 and 3 the same.

r=kurt

git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@21549 5aad4ff5-4b58-4159-ba3f-0e6328df4cad