Allow registering of internal search engines
This changes gets rid of the non-standard search engine tags to
determine the function of a searchengine. Instead a searchengine has to
be registered explicitly to be internal with the function
window.registerSearchEngineHandler(searchEngineAlias[, liveSearch]).
Engines can also be unregistered. The songbirdlibrary search is
registered by default. This is for of #144.