components/search, components/remoteapi updates, not sure why the modification to autodefs.mk.in fixed the idl compilations, too tired to think about it...
But... the whole tree compiles now...
No promises on functionality haha.
Note to self: go back and check previous *** commits and replace
reentrantmutexes with mutexes

The new deps I built still didn't export libmozjs.so, so I'll have to
tweak that, but it did compile it. Dropping that into
mozilla-9.0.1/debug/lib works for now.

link to libmozjs.so
https://dl.dropboxusercontent.com/u/26254886/libmozjs.so-debug
(need to rename it to libmozjs.so when putting it into the deps dir, for
some reason the release dep build didn't compile it.)